RDM-4253: Access Control (#194)
* Switch to ES6 imports for all routes

Use ES6 module imports in place of CommonJS requires.

* RDM-4013 Restrict visibility of Manage User Profiles link in CCD Admin Web

* RDM-4013 Restict access to manage user profile

* RDM-4003: Move Import Definition page from initial screen

Move the Import Definition functionality onto its own page, off the Home page. Add basic placeholder page for Home screen.

* interim commit before merge from master

* Resolve conflicts

* Resolve conflicts

* Remove file that were not in repo

* RDM-4013 interim commit

* RDM-4013 added test coverage

* RDM-4013 add coverage

* RDM-4013

* RDM-4013 revert

* RDM-4013 add terraform configs

* correct type:

* Resolve conflicts

* RDM-4253 Access control

* RDM-4253 code tidy up

* Fix linting error

* Fix test failures

+++ src/main/routes/deleteDefinition.ts
@@ -1,11 +1,14 @@
 import { deleteDefinition } from "../service/delete-definition-service";
 import { deleteSessionVariables } from "../util/clearSession";
+import { error_unauthorized_role } from "../util/error_unauthorized_role";
 import router from "./home";
 
+const errorPage = "error";
 /**
  * Delete Definition based on jurisdiction ID and version
  */
 router.post("/deletedefinition", (req, res, next) => {
+  if (req.adminWebAuthorization && req.adminWebAuthorization.canManageDefinition) {
     if (req.body.deleteItem === "Yes") {
         deleteDefinition(req).then((response) => {
             deleteSessionVariables(req);
@@ -28,6 +31,9 @@ router.post("/deletedefinition", (req, res, next) => {
         res.redirect(302, `/deleteitem?item=${req.body.itemToDelete}&jurisdictionId=${req.body.jurisdictionId}`
           + `&version=${req.body.definitionVersion}`);
     }
+  } else {
+    res.render(errorPage, error_unauthorized_role());
+  }
 });
 
 export default router;
